+++ src/ArubapecClient.php
@@ -7,6 +7,7 @@ namespace Shellrent\Arubapec;
 use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 use Shellrent\Arubapec\Account\AccountClient;
+use Shellrent\Arubapec\AdditionalService\AdditionalServiceClient;
 use Shellrent\Arubapec\Auth\AuthClient;
 use Shellrent\Arubapec\Config\ClientConfig;
 
@@ -19,6 +20,8 @@ final class ArubapecClient
     private readonly AuthClient $authClient;
 
     private readonly AccountClient $accountClient;
+
+    private readonly AdditionalServiceClient $additionalServiceClient;
 
     public function __construct(
         ?ClientInterface $httpClient = null,
@@ -33,6 +36,7 @@ final class ArubapecClient
         $this->httpClient = $httpClient ?? new Client($this->config->getHttpClientConfig());
         $this->authClient = new AuthClient($this->httpClient);
         $this->accountClient = new AccountClient($this->httpClient);
+        $this->additionalServiceClient = new AdditionalServiceClient($this->httpClient);
     }
 
     public function auth(): AuthClient
@@ -43,6 +47,11 @@ final class ArubapecClient
     public function account(): AccountClient
     {
         return $this->accountClient;
+    }
+
+    public function additionalService(): AdditionalServiceClient
+    {
+        return $this->additionalServiceClient;
     }
 
     public function getHttpClient(): ClientInterface