+++ tests/Account/AccountClientTest.php
@@ -32,7 +32,7 @@ use Shellrent\Arubapec\Account\Dto\AccountTypesResponse;
 use Shellrent\Arubapec\Account\Dto\Interval;
 use Shellrent\Arubapec\Account\Dto\OwnerId;
 use Shellrent\Arubapec\Account\Dto\OwnerSearchRequest;
-use Shellrent\Arubapec\Account\Dto\RenewalData;
+use Shellrent\Arubapec\Shared\Dto\RenewalData;
 use Shellrent\Arubapec\Exception\ApiException;
 
 final class AccountClientTest extends TestCase